% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breusch_pagan_test.R
\name{breusch_pagan_test}
\alias{breusch_pagan_test}
\title{Breusch-Pagan Test for Heteroscedasticity}
\usage{
breusch_pagan_test(data, model_results)
}
\arguments{
\item{data}{A data frame where the first column is the dependent variable (Y) and the remaining columns are the independent variables (X1, X2, ...).}

\item{model_results}{A list containing the results from the `linear_model` function.}
}
\value{
A list containing the Breusch-Pagan test statistic and the corresponding p-value.
}
\description{
Breusch-Pagan Test for Heteroscedasticity
}
