% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{ginv}
\alias{ginv}
\title{Generalized Inverse of a Matrix (Moore-Penrose)}
\usage{
ginv(A, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{A}{A matrix to be inverted.}
}
\value{
The Moore-Penrose pseudo-inverse of matrix A.
}
\description{
Generalized Inverse of a Matrix (Moore-Penrose)
}
