% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_se_hc2.R
\name{robust_se_hc2}
\alias{robust_se_hc2}
\title{Robust Standard Errors (HC2)}
\usage{
robust_se_hc2(model_results)
}
\arguments{
\item{model_results}{A list containing the results from the `linear_model` function.}
}
\value{
A data frame containing the coefficients, robust standard errors, and p-values.
}
\description{
Robust Standard Errors (HC2)
}
