% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_se_hc3.R
\name{robust_se_hc3}
\alias{robust_se_hc3}
\title{Calculate Robust Standard Errors (HC3)}
\usage{
robust_se_hc3(data, model_results)
}
\arguments{
\item{data}{A data frame where the first column is the dependent variable (Y) and the remaining columns are the independent variables (X1, X2, ...).}

\item{model_results}{A list containing the results from the `linear_model` function.}
}
\value{
A list containing the estimated coefficients, robust standard errors, p-values, adjusted R-squared, F-statistic, and p-value of the F-test.
}
\description{
Calculate Robust Standard Errors (HC3)
}
