% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_test.R
\name{white_test}
\alias{white_test}
\title{White Test for Heteroscedasticity}
\usage{
white_test(data, model_results)
}
\arguments{
\item{data}{A data frame where the first column is the dependent variable (Y) and the remaining columns are the independent variables (X1, X2, ...).}

\item{model_results}{A list containing the results from the `linear_model` function.}
}
\value{
A list containing the White test statistic and the corresponding p-value.
}
\description{
White Test for Heteroscedasticity
}
